moved integration tests under test_version_store.py where they belong
